requirements.txt
bcrypt==4.1.1
click==8.1.7
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.0
greenlet==3.0.2
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.28
protobuf==4.25.1
SQLAlchemy==2.0.23
typing-extensions==4.9.0
Werkzeug==2.2.2
wtforms==3.1.1
zipp==3.17.0
